+++ src/router/routes.js
@@ -7,7 +7,7 @@ import SignUp from '../views/SignUp.vue'
 import SocialView from '../views/SocialView.vue'
 import SettingsView from '../views/SettingsView.vue'
 import SinglePostView from '../views/SinglePostView.vue'
-import NotFound from '../views/NotFound.vue'
+import NotFoundView from '../views/NotFoundView.vue'
 
 // Routes go here
 export default [
@@ -61,7 +61,11 @@ export default [
       window.location.assign('/404') // redirect to 404 page to trigger Django to send 404 response
     },
     name: 'NotFound',
-    component: NotFound
+    component: NotFoundView
   },
-  { path: '/404', name: '404', component: NotFound }
+  {
+    path: '/404',
+    name: '404',
+    component: NotFoundView
+  }
 ]
