+++ src/views/NotFoundView.vue
